refactor: move TDBID extension in Utils
