+++ src/rtti.cpp
@@ -16,7 +16,6 @@
 #include "App/Callback/Targets/InputTarget.hpp"
 #include "App/Callback/Targets/ResourceTarget.hpp"
 #include "App/Callback/Targets/StaticEntityTarget.hpp"
-#include "App/Data/TDBIDHelperEx.hpp"
 #include "App/Depot/CResourceEx.hpp"
 #include "App/Depot/CurveData.hpp"
 #include "App/Depot/MeshEx.hpp"
@@ -69,6 +68,7 @@
 #include "App/Utils/NodeRef.hpp"
 #include "App/Utils/Number.hpp"
 #include "App/Utils/String.hpp"
+#include "App/Utils/TDBIDHelperEx.hpp"
 #include "App/World/DynamicEntityEvent.hpp"
 #include "App/World/DynamicEntitySpec.hpp"
 #include "App/World/DynamicEntityState.hpp"
